+++ test/integration/getVersion.test.ts
@@ -15,9 +15,14 @@ describe("getVersion", async () => {
   });
 
   it("Should correctly get the semantic version of device", async () => {
-    const { version } = await ada.getVersion();
+    const { version, compatibility } = await ada.getVersion();
 
     expect(version.major).to.equal(2);
     expect(version.minor).to.equal(2);
+    expect(compatibility).to.deep.equal({
+      isCompatible: true,
+      recommendedVersion: null,
+      supportsMary: true,
+    })
   });
 });